Improvements for my amazing population mapping pipeline

+++ src/toil_vg/test/test_vg.py
@@ -49,7 +49,11 @@ class VGCGLTest(TestCase):
             shutil.copyfileobj(connection, f)
 
     def setUp(self):
-        self.workdir = tempfile.mkdtemp()
+        # Set this to True to poke around in the outsores for debug purposes
+        self.saveWorkDir = False
+        self.workdir = './toil-vgci_work' if self.saveWorkDir else tempfile.mkdtemp()
+        if not os.path.exists(self.workdir):
+            os.makedirs(self.workdir)
         self.jobStoreLocal = '{}/local-testvg-{}'.format(self.workdir, uuid4())
 
         # input files all in same bucket folder, which is specified (only) here:
@@ -83,13 +87,13 @@ class VGCGLTest(TestCase):
         self.baseline = self._ci_input_path('small.vcf.gz')
         self.chrom_fa = self._ci_input_path('small.fa.gz')
 
-        self._run(self.base_command, self.jobStoreLocal, 'sample',
+        self._run(self.base_command, self.jobStoreLocal, '1',
                   self.local_outstore,  '--fastq', self.sample_reads,
                   '--graphs', self.test_vg_graph,
                   '--chroms', 'x', '--vcfeval_baseline', self.baseline,
                   '--vcfeval_fasta', self.chrom_fa, '--vcfeval_opts', ' --squash-ploidy')
         
-        self._assertOutput('sample', self.local_outstore, f1_threshold=0.95)
+        self._assertOutput('1', self.local_outstore, f1_threshold=0.95)
 
     def test_02_sim_small_standalone(self):
         ''' 
@@ -192,9 +196,10 @@ class VGCGLTest(TestCase):
                   '--gams', os.path.join(self.local_outstore, 'aligned-vg_default.gam'),
                    os.path.join(self.local_outstore, 'aligned-vg-pe_default.gam'),
                   '--gam_names', 'vg', 'vg-pe',
+                  '--realTimeLogging', '--logInfo',
                   '--vcfeval_fasta', self.chrom_fa_nz,
                   '--vcfeval_baseline', self.baseline,
-                  '--sample_name', 'x',
+                  '--sample_name', '1',
                   '--calling_cores', '2',
                   '--genotype', '--genotype_opts', '-P 0', 
                   '--freebayes',
@@ -337,14 +342,14 @@ class VGCGLTest(TestCase):
         self.baseline = self._ci_input_path('small.vcf.gz')
         self.chrom_fa = self._ci_input_path('small.fa.gz')
 
-        self._run(self.base_command, self.jobStoreLocal, 'sample',
+        self._run(self.base_command, self.jobStoreLocal, '1',
                   self.local_outstore,  '--fastq', self.sample_reads,
                   '--graphs', self.test_vg_graph,
                   '--chroms', 'x', '--vcfeval_baseline', self.baseline,
                   '--vcfeval_fasta', self.chrom_fa, '--vcfeval_opts', ' --squash-ploidy',
                   '--force_outstore')
 
-        self._assertOutput('sample', self.local_outstore, f1_threshold=0.95)
+        self._assertOutput('1', self.local_outstore, f1_threshold=0.95)
 
     def test_07_construct(self):
         '''
@@ -400,14 +405,14 @@ class VGCGLTest(TestCase):
         self.baseline = self._ci_input_path('small.vcf.gz')
         self.chrom_fa = self._ci_input_path('small.fa.gz')
 
-        self._run(self.base_command, self.jobStoreLocal, 'sample',
+        self._run(self.base_command, self.jobStoreLocal, '1',
                   self.local_outstore,  '--fastq', self.sample_reads,
                   '--graphs', self.test_vg_graph,
                   '--chroms', 'x', '--vcfeval_baseline', self.baseline,
                   '--vcfeval_fasta', self.chrom_fa, '--vcfeval_opts', ' --squash-ploidy',
                   '--genotype', '--genotype_opts', ' -Q -A')
 
-        self._assertOutput('sample', self.local_outstore, f1_threshold=0.95)
+        self._assertOutput('1', self.local_outstore, f1_threshold=0.95)
 
     def test_09_sim_small_genotype_no_augment(self):
         ''' 
@@ -418,14 +423,14 @@ class VGCGLTest(TestCase):
         self.baseline = self._ci_input_path('small.vcf.gz')
         self.chrom_fa = self._ci_input_path('small.fa.gz')
 
-        self._run(self.base_command, self.jobStoreLocal, 'sample',
+        self._run(self.base_command, self.jobStoreLocal, '1',
                   self.local_outstore,  '--fastq', self.sample_reads,
                   '--graphs', self.test_vg_graph,
                   '--chroms', 'x', '--vcfeval_baseline', self.baseline,
                   '--vcfeval_fasta', self.chrom_fa, '--vcfeval_opts', ' --squash-ploidy',
                   '--genotype', '--no_augment', '--genotype_opts', ' -Q -A')
 
-        self._assertOutput('sample', self.local_outstore, f1_threshold=0.95)
+        self._assertOutput('1', self.local_outstore, f1_threshold=0.95)
 
     def test_10_gbwt(self):
         '''
@@ -514,6 +519,7 @@ class VGCGLTest(TestCase):
         self.assertEqual(headers, set(names))
         
     def tearDown(self):
-        shutil.rmtree(self.workdir)
+        if not self.saveWorkDir:
+            shutil.rmtree(self.workdir)
         subprocess.check_call(['toil', 'clean', self.jobStoreLocal])
